Working on recursing configuration for configured properties. This is a checkpoint commit as there is some cleanup to do.
e.g. scheduled methods are have an intermediary representation plus extra configurations like schedules and timeouts etc.

The old concept of key builders is gone and the keys are built up by calling join(delimiter) on the configured thing which gives the unique key, turns out this is very useful for mapping to
various configuration sources. Given a key of (a,b,c) it can be a.b.c or A_B_C or --a-b-c depending if your source is system, environment or command line.