Allow for inverted configuration values
Inverted values are source from the DI container. In order to support nested configurations, when a value is resolved its registered
with the container and inverted. A second pass allows discovered configured fields to initialised.

This change allows for values that are sources from the container and as such do not need inversion.